Cleaned up VS2022 project and added proper x64 / ARM64 / ARM64X targets

+++ lib/regcomp.c
@@ -57,7 +57,7 @@ tre_regncomp(regex_t *preg, const char *regex, size_t n, int cflags)
 #endif /* HAVE_MBSTATE_T */
       while (n > 0)
 	{
-	  consumed = tre_mbrtowc(wcptr, regex, n, &state);
+	  consumed = (int)tre_mbrtowc(wcptr, regex, n, &state);
 
 	  switch (consumed)
 	    {
@@ -77,7 +77,7 @@ tre_regncomp(regex_t *preg, const char *regex, size_t n, int cflags)
 	    case -2:
 	      /* The last character wasn't complete.  Let's not call it a
 		 fatal error. */
-	      consumed = n;
+	      consumed = (int)n;
 	      break;
 	    }
 	  regex += consumed;
